add admin functions (fix ui, genre/interests)

+++ backend/src/main/java/co/movielearn/interests/InterestRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface InterestRepository extends JpaRepository<Interest, Integer> {
+public interface InterestRepository extends JpaRepository<Interest, Long> {
 
     Optional<Interest> findByName(String name);
